refactor: 프로젝트 구조 변경
- exception, constants, message를 common 패키지로 묶음.
