+++ src/main/java/lotto/common/constants/Constants.java
@@ -1,4 +1,4 @@
-package lotto.constants;
+package lotto.common.constants;
 
 public class Constants {
