+++ src/test/java/lotto/util/ParserTest.java
@@ -6,8 +6,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
-import lotto.exception.LottoException;
-import lotto.message.ErrorMessage;
+import lotto.common.exception.LottoException;
+import lotto.common.message.ErrorMessage;
 import lotto.vo.Budget;
 import lotto.vo.DrawnLottoNumber;
 import lotto.vo.lotto.Lotto;
